pyproject.toml
[tool.poetry]
name = "bear-note-graph"
version = "0.1.0"
description = "Generate a graphviz visualisation of your Bear notes. It also has a partial Markdown parser, because why not"
authors = ["Ruben Berenguel <ruben+poetry@mostlymaths.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
colorlog = "^4.1.0"
pyyaml = "^5.3"
mypy = "^0.770"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4.4"
flake8 = "^3.7.9"
isort = "^4.3.21"
black = "^19.10b0"
mypy = "^0.770"
monkeytype = "^19.11.2"

[tool.poetry.scripts]
bear-note-graph = 'bear_note_graph.main:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

